requirements.txt
requests
pillow
tqdm
pandas